k1_map: more efficient key and value for sys_auth_map
Previously sys_auth_map used uid as key, it wasn't very good as we had to search for the exact auth_type in the registered list for a uid. The new key is a struct that has auth_type and uid together.
This change implies that we don't need to have the auth_type in our record, nor we need to have a list as the value for sys_auth_map, sys_record is simplified and now only contains 1. verdict cache, 2. destination verdict type and 3. credential for the current auth_record.